back-end/requirement.txt
PyMySQL==1.0.2
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Limiter==2.2.0
PyJWT==2.3.0
gunicorn==20.1.0